python/requirements.txt
flask==3.0.0
sqlalchemy==2.0.25
jinja2==3.1.3
astroid==3.0.2
libcst==1.8.6
parso==0.8.3
google-generativeai==0.3.2
pyyaml==6.0.3
black==24.1.1
click==8.1.7
pytest==7.4.4
